refactor: saveReview -> create 메서드명 변경 및 command 사용 (#316)
