--- backend/src/main/java/com/celuveat/restaurant/application/dto/SaveReviewRequest.java
@@ -1,6 +1,0 @@
-package com.celuveat.restaurant.application.dto;
-
-public record SaveReviewRequest(
-        String content
-) {
-}
