+++ backend/src/main/java/com/celuveat/restaurant/application/dto/SaveReviewRequestCommand.java
@@ -1,0 +1,8 @@
+package com.celuveat.restaurant.application.dto;
+
+public record SaveReviewRequestCommand(
+        String content,
+        Long memberId,
+        Long restaurantId
+) {
+}
